LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/selectedcontent-mutations-expected.txt









FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test1 assert_equals: test1 expected "<select>\n    <button>\n      <selectedcontent><span>span</span> three</selectedcontent>\n    </button>\n    <option><span>span</span> one</option>\n    <option><span>span</span> two</option>\n    <option selected=\"\"><span>span</span> three</option>\n  </select>" but got "<select>\n    <button>\n      <selectedcontent></selectedcontent>\n    </button>\n    <option><span>span</span> one</option>\n    <option><span>span</span> two</option>\n    <option selected=\"\"><span>span</span> three</option>\n  </select>"
FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test2 assert_equals: test2 expected "<select>\n    <selectedcontent>option</selectedcontent>\n    <option>option</option>\n  </select>" but got "<select>\n    <selectedcontent></selectedcontent>\n    <option>option</option>\n  </select>"
FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test3 assert_equals: test3 expected "<select>\n    <button>\n      <selectedcontent>option</selectedcontent>\n      <selectedcontent>outer2\n        <selectedcontent>inner2</selectedcontent>\n      </selectedcontent>\n    </button>\n    <option>option</option>\n  </select>" but got "<select>\n    <button>\n      <selectedcontent>outer1\n        <selectedcontent>inner1</selectedcontent>\n      </selectedcontent>\n      <selectedcontent>outer2\n        <selectedcontent>inner2</selectedcontent>\n      </selectedcontent>\n    </button>\n    <option>option</option>\n  </select>"
FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test4 assert_equals: test4 expected "<select>\n    <button>\n      <selectedcontent>two</selectedcontent>\n    </button>\n    <div>\n      <option>one</option>\n    </div>\n    <div>\n      <option selected=\"\">two</option>\n    </div>\n  </select>" but got "<select>\n    <button>\n      <selectedcontent></selectedcontent>\n    </button>\n    <div>\n      <option>one</option>\n    </div>\n    <div>\n      <option selected=\"\">two</option>\n    </div>\n  </select>"
FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test5 assert_equals: test5 expected "<select>\n    <option>option</option>\n    <button>\n      <selectedcontent>option</selectedcontent>\n    </button>\n  </select>" but got "<select>\n    <option>option</option>\n    <button>\n      <selectedcontent></selectedcontent>\n    </button>\n  </select>"
PASS MutationObserver records during parsing of <select> with <selectedcontent>: test6
FAIL MutationObserver records during parsing of <select> with <selectedcontent>: test7 assert_equals: test7 expected "<select>\n    <button>\n      <selectedcontent>one</selectedcontent>\n    </button>\n    <option>one</option>\n    <div></div>\n  <option selected=\"\">two</option><option selected=\"\">three</option></select>" but got "<select>\n    <button>\n      <selectedcontent></selectedcontent>\n    </button>\n    <option>one</option>\n    <div></div>\n  <option selected=\"\">two</option><option selected=\"\">three</option></select>"
